Return children from forgetChildren instead of calling maybeReportOrphan.
If the caller of forgetChildren wishes to report the forgotten children as
potential orphans, it can simply call .forEach(maybeReportOrphan).

The real win here is being able to wait until after the recomputation
finishes in reallyRecompute to report any orphans, since often the
recomputation results in re-remembering the forgotten children.