Rewrite Entry class to use module functions for private methods.
Because the function names can now be minified to single characters, this
results in significant bundle size savings (17.5% minified, 5.7% after
gzip) and a modest performance improvement.

Making all Entry properties public (so that they can be consumed by
external functions) is not a problem because the entire Entry class is a
private API of the optimism package.