Cache exceptions as well as ordinary values.
Previously, if a wrapped function threw an exception, no result would be
cached, and the function would be recomputed the next time its wrapper was
called, even if nothing upon which it depended had changed.

One of the goals of this library is to work equally well with Promise
results as it does with ordinary results. Since a Promise can represent
either an ordinary value or an exception, it seems important for the
non-Promise behavior of this library to align with Promise behavior.

This means, most importantly, that we should be caching exceptions as well
as ordinary results, to avoid doing unnecessary work just because the
previous invocation of a cached function threw an exception.

If you call a cached function multiple times and get the same (===)
rejected Promise object, the rejection values will be === to each other,
because a Promise can only have one rejection value. The same should be
true for a function that throws an Error object (or any JavaScript value,
for that matter): thrown values should be === to one another if nothing
has changed, just as ordinary results are === to one another.

This new exception-caching behavior is definitely a minor breaking change,
but you can reproduce the old semantics by calling wrapper.dirty(...args)
if an exception is thrown:

  try {
    result = cachedWrapperFunction(a, b);
  } catch (e) {
    cachedWrapperFunction.dirty(a, b);
    throw e;
  }

Internally, unknown, ordinary, and exceptional values are represented as
arrays: [], [value], and [, exception], respectively.