Support options.keyArgs to complement options.makeCacheKey. (#65)
This optional function takes the raw arguments passed to an
OptimisticFunctionWrapper and simplifies/filters/reduces them down to an
array of arguments sufficient to identify the cache result. This
simplified array of arguments is then passed to options.makeCacheKey (or
defaultMakeCacheKey if no options.makeCacheKey was provided).

The resulting array serves as the arguments list for any function that
only needs to compute a cache key (and thus does not need the full raw
arguments), such as wrapper.dirty and options.makeCacheKey. This makes it
possible to dirty a cached function using less information than would be
required to call it properly, which is useful in some cases.

In some sense, options.keyArgs and options.makeCacheKey divide up the work
of computing cache keys from raw arguments. Both functions are optional,
and any combination of the two that satisfies the type system is
allowed. If the sequence of arguments returned by options.keyArgs can be
passed directly to defaultMakeCacheKey, you may not need to define a
custom options.makeCacheKey function anymore.